requirements.txt
asn1crypto==0.24.0
cffi==1.11.2
cryptography==2.1.4
Django==2.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-webpack-loader==0.5.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
ecdsa==0.13
future==0.16.0
idna==2.6
pkg-resources==0.0.0
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.5.3
python-jose==1.4.0
pytz==2017.3
six==1.11.0
